requirements.txt
flask==3.1.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
slack-sdk==3.34.0
anthropic==0.52.0
python-dotenv==1.1.0
